* Se agregan archivos faltantes del feature de los Servidores automágicos * Se agregan canales de comunicación entre Despacho/Robot16 y Cliente. Hay que refactorizar y discutir   ciertas partes del diseño del mismo. Se probó el mismo y, al menos, en la última prueba realizada   se logró que un cliente ingresara a la planta y se fuera con su pedido satisfecho.
POR HACER:
* No se matan los canales de comunicación Despacho-Cliente / R16-Cliente al irse un cliente. Habria que 
  analizar como terminar elegantemente esta comunicación.
* Distribuir al despacho.

+++ trunk/distribuidos/EjGrupal/branches/Iteracion_1/src/Comunicaciones/CanalEntradaCliente.cpp
@@ -17,13 +17,13 @@
 int main(int argc, char **argv) {
     CommunicationsUtil util;
     char buffer[255];
-    int nroCliente;
+    long nroCliente;
 
     if ( util.parseArgs(argc, argv, nroCliente) == -1 ) {
         exit(1);
     }
 
-    sprintf(buffer, "Canal Entrada Cliente %d:", nroCliente);
+    sprintf(buffer, "Canal Entrada Cliente %ld:", nroCliente);
     Logger::getInstance().setProcessInformation(buffer);
     Logger::logMessage(Logger::COMM, "Conectando canal de entrada");
